+++ src/FluxFlow.Components.Storage/Contracts/IStorageStoreHandle.cs
@@ -1,0 +1,6 @@
+namespace FluxFlow.Components.Storage.Contracts;
+
+public interface IStorageStoreHandle
+{
+    string StoreName { get; }
+}