+++ src/FluxFlow.Components.Storage/StorageComponentTypes.cs
@@ -4,6 +4,7 @@ namespace FluxFlow.Components.Storage;
 
 public static class StorageComponentTypes
 {
+    public static readonly NodeType Store = new("storage.store");
     public static readonly NodeType Put = new("storage.put");
     public static readonly NodeType Get = new("storage.get");
     public static readonly NodeType Query = new("storage.query");